+++ spring-cloud-starter-tencent-polaris-router/src/main/java/com/tencent/cloud/polaris/router/interceptor/MetadataRouterRequestInterceptor.java
@@ -21,6 +21,7 @@ package com.tencent.cloud.polaris.router.interceptor;
 import java.util.Map;
 import java.util.Set;
 
+import com.tencent.cloud.common.constant.RouterConstant;
 import com.tencent.cloud.polaris.router.PolarisRouterContext;
 import com.tencent.cloud.polaris.router.config.properties.PolarisMetadataRouterProperties;
 import com.tencent.cloud.polaris.router.spi.RouterRequestInterceptor;
@@ -29,7 +30,7 @@ import com.tencent.polaris.router.api.rpc.ProcessRoutersRequest;
 
 /**
  * Router request interceptor for metadata router.
- * @author lepdou 2022-07-06
+ * @author lepdou, Hoatian Zhang
  */
 public class MetadataRouterRequestInterceptor implements RouterRequestInterceptor {
 	private static final String LABEL_KEY_METADATA_ROUTER_KEYS = "system-metadata-router-keys";
@@ -49,7 +50,7 @@ public class MetadataRouterRequestInterceptor implements RouterRequestIntercepto
 		// 1. get metadata router label keys
 		Set<String> metadataRouterKeys = routerContext.getLabelAsSet(LABEL_KEY_METADATA_ROUTER_KEYS);
 		// 2. get metadata router labels
-		Map<String, String> metadataRouterLabels = routerContext.getLabels(PolarisRouterContext.ROUTER_LABELS,
+		Map<String, String> metadataRouterLabels = routerContext.getLabels(RouterConstant.ROUTER_LABELS,
 				metadataRouterKeys);
 		// 3. set metadata router labels to request
 		request.addRouterMetadata(MetadataRouter.ROUTER_TYPE_METADATA, metadataRouterLabels);