+++ spring-cloud-starter-tencent-polaris-router/src/main/java/com/tencent/cloud/polaris/router/scg/PolarisReactiveLoadBalancerClientFilter.java
@@ -27,7 +27,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import com.tencent.cloud.common.constant.RouterConstants;
+import com.tencent.cloud.common.constant.RouterConstant;
 import com.tencent.cloud.common.metadata.MetadataContext;
 import com.tencent.cloud.common.metadata.MetadataContextHolder;
 import com.tencent.cloud.common.metadata.StaticMetadataManager;
@@ -78,7 +78,7 @@ import static org.springframework.cloud.gateway.support.ServerWebExchangeUtils.a
  * with PolarisReactiveLoadBalancerClientFilter. The passed route labels are used in
  * {@link PolarisRouterServiceInstanceListSupplier}.
  *
- * @author lepdou 2022-06-20
+ * @author lepdou, Hoatian Zhang
  */
 public class PolarisReactiveLoadBalancerClientFilter extends ReactiveLoadBalancerClientFilter {
 	private static final Logger log = LoggerFactory.getLogger(PolarisReactiveLoadBalancerClientFilter.class);
@@ -210,7 +210,7 @@ public class PolarisReactiveLoadBalancerClientFilter extends ReactiveLoadBalance
 	// the router label is passed through the http header uniformly instead of the original hint mechanism.
 	HttpHeaders genRouterHttpHeaders(ServerWebExchange exchange, String peerServiceName) {
 		HttpHeaders headers = new HttpHeaders();
-		headers.add(RouterConstants.ROUTER_LABEL_HEADER, genRouterHint(exchange, peerServiceName));
+		headers.add(RouterConstant.ROUTER_LABEL_HEADER, genRouterHint(exchange, peerServiceName));
 		return headers;
 	}